fix ci: fix json related failures
With failures like
```
<3>[filter_expr_executor.cpp:167] Type for key Count was 6 which does not have a comparison operator
<3>[filter_expr_executor.cpp:356] Fell through.  Should never happen.  Attempt to compare type 2 to type 0
<3>[filter_expr_executor.cpp:167] Type for key Count was 6 which does not have a comparison operator
<3>[filter_expr_executor.cpp:356] Fell through.  Should never happen.  Attempt to compare type 0 to type 2
../test/redfish-core/include/filter_expr_executor_test.cpp:40: Failure
Expected equality of these values:
  json["Members"].size()
    Which is: 1
  0
```

The issue is that the integer literal is mapped to an unsigned int type.

For this type the comparison was not implemented and it fell through.

For failures like

```
[----------] Global test environment tear-down
[==========] 34 tests from 9 test suites ran. (0 ms total)
[  PASSED  ] 29 tests.
[  FAILED  ] 5 tests, listed below:
[  FAILED  ] ReadJson.JsonSubElementValueAreUnpackedCorrectly
[  FAILED  ] ReadJson.MultipleJsonSubElementValueAreUnpackedCorrectly
[  FAILED  ] ReadJsonPatch.ValidElementsReturnsTrueResponseOkValuesUnpackedCorrectly
[  FAILED  ] ReadJsonPatch.OdataIgnored
[  FAILED  ] ReadJsonAction.ValidElementsReturnsTrueResponseOkValuesUnpackedCorrectly
```

the issue was that it was mapped to an uint64_t and the type had to be
adjusted.

Note only the tests have been adjusted and there may be a need to make
adjustments in the other callers as well.

The failures were likely introduced by the changes in
openbmc-build-scripts commit e4b761fe0a9c789c2c8a7147081c138ad9e19268
but did not confirm this.

The upgrade made changes to boost and nlohmann json versions.
```
boost
-        rev="1.87.0",
+        rev="1.88.0",
nlohmann/json
-        rev="v3.11.3",
+        rev="v3.12.0",
```

Tested: Unit Tests Pass. Needs expert review.

+++ test/redfish-core/include/utils/json_utils_test.cpp
@@ -182,7 +182,7 @@ TEST(ReadJson, JsonSubElementValueAreUnpackedCorrectly)
         }
     )"_json;
 
-    int integer = 0;
+    uint64_t integer = 0;
     ASSERT_TRUE(readJson(jsonRequest, res, "json/integer", integer));
     EXPECT_EQ(integer, 42);
     EXPECT_EQ(res.result(), boost::beast::http::status::ok);
@@ -220,7 +220,7 @@ TEST(ReadJson, MultipleJsonSubElementValueAreUnpackedCorrectly)
         }
     )"_json;
 
-    int integer = 0;
+    uint64_t integer = 0;
     std::string foobar;
     std::string bazbar;
     ASSERT_TRUE(readJson(jsonRequest, res, "json/integer", integer,
@@ -325,7 +325,7 @@ TEST(ReadJsonPatch, ValidElementsReturnsTrueResponseOkValuesUnpackedCorrectly)
     // Ignore errors intentionally
     req.addHeader(boost::beast::http::field::content_type, "application/json");
 
-    int64_t integer = 0;
+    uint64_t integer = 0;
     ASSERT_TRUE(readJsonPatch(req, res, "integer", integer));
     EXPECT_EQ(res.result(), boost::beast::http::status::ok);
     EXPECT_THAT(res.jsonValue, IsEmpty());
@@ -353,7 +353,7 @@ TEST(ReadJsonPatch, OdataIgnored)
     req.addHeader(boost::beast::http::field::content_type, "application/json");
     // Ignore errors intentionally
 
-    std::optional<int64_t> integer = 0;
+    std::optional<uint64_t> integer = 0;
     ASSERT_TRUE(readJsonPatch(req, res, "integer", integer));
     EXPECT_EQ(res.result(), boost::beast::http::status::ok);
     EXPECT_THAT(res.jsonValue, IsEmpty());
@@ -418,7 +418,7 @@ TEST(ReadJsonAction, ValidElementsReturnsTrueResponseOkValuesUnpackedCorrectly)
     req.addHeader(boost::beast::http::field::content_type, "application/json");
     // Ignore errors intentionally
 
-    int64_t integer = 0;
+    uint64_t integer = 0;
     ASSERT_TRUE(readJsonAction(req, res, "integer", integer));
     EXPECT_EQ(res.result(), boost::beast::http::status::ok);
     EXPECT_THAT(res.jsonValue, IsEmpty());
